% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{AddFactor}
\alias{AddFactor}
\title{Utilities Add Factors}
\usage{
AddFactor(object, colData = NULL, rowData = NULL, value = NULL)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{colData}{Input the names that will be added into coldata of RISC object, 
it should be characters, as the col.names of coldata.}

\item{rowData}{Input the names that will be added into rowdata of RISC object,
it should be characters, as the col.names of rowdata.}

\item{value}{The factor vector or data.frame that will be added into coldata 
or rowdata, the vector/data.frame should have equal names/row.names to the 
row.names coldata or rowdata of RISC object. The input: vector or data.frame.}
}
\description{
The "AddFactor" function can add factors to the full dataset, this function 
can add one or more factors into coldata. Here the row.names/names of factor 
matrix/vector should be equal to the row.names of coldata of RISC object.
}
