% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrating.R
\name{scAlignGene}
\alias{scAlignGene}
\title{Integrating Multiple Datasets}
\usage{
scAlignGene(
  objects,
  Align = "OLS",
  ncore = 1,
  Add.Id = NULL,
  do.fast = "AUTO",
  beta = NULL,
  seed = 123
)
}
\arguments{
\item{objects}{The list of multiple RISC objects: 
list{object1, object2, object3, ...}.}

\item{Align}{The method for alignment of gene expression values: "Optimal" for 
alignment by experience, "Predict" for alignment by RPCI prediction, and "OLS" 
for alignment by the ordinary linear regression.}

\item{ncore}{The number of multiple cores for data integration.}

\item{Add.Id}{Add a vector of Id to label different datasets, a character vector.}

\item{do.fast}{With enough PC memory, we could do.fast, otherwise, let it as the 
default. Three options: AUTO (default), TRUE, and FALSE.}

\item{beta}{The coefficiency to generate predicted counts.}

\item{seed}{The random seed to keep consistent result.}
}
\description{
The "scMultiIntegrate" function can be used for data integration of multiple 
datasets, it is basically based on our new algorithm reference principal 
components integration (RPCI). RPCI decomposes all the target datasets based 
on the reference data. The output of this function aligns the gene expression 
of different datasets.
}
\references{
Liu et al., Nature Biotech. (2021)
}
