% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMarker.R
\name{AllMarker}
\alias{AllMarker}
\title{Find All Cluster Markers}
\usage{
AllMarker(
  object,
  positive = TRUE,
  frac = 0.25,
  logFC = 0.5,
  Padj = 0.05,
  latent.factor = NULL,
  min.cells = 25L,
  method = "NB",
  ncore = 1
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{positive}{Whether only output the cluster markers with positive logFC.}

\item{frac}{A fraction cutoff, the cluster marker genes expressed at least a 
cutoff fraction of the cluster cells.}

\item{logFC}{The cutoff of log Fold-change for differentially expressed marker 
genes.}

\item{Padj}{The cutoff of the adjusted P-value.}

\item{latent.factor}{The latent factor from coldata, which represents number 
values or factors, and only one latent factor can be inputed.}

\item{min.cells}{The threshold for the cell number of valid clusters.}

\item{method}{Which method is used to identify cluster markers, two options: 
'NB' for Negative Binomial model and 'QP' for QuasiPoisson model.}

\item{ncore}{The multiple cores for parallel calculating.}
}
\description{
This function depends on "scMarker" function by using the same criteria, and 
generates markers for all the clusters. Here, if the cell number of any cluster 
is less than 10 , RISC will skip this cluster and not detect its cluster markers 
in the default parameters.
}
