% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMarker.R
\name{scDEG}
\alias{scDEG}
\title{Find Differentially Expressed Genes between Clusters}
\usage{
scDEG(
  object,
  cluster.ctrl = NULL,
  cluster.sam = NULL,
  cell.ctrl = NULL,
  cell.sam = NULL,
  frac = 0.1,
  logFC = 0.25,
  Padj = 0.05,
  latent.factor = NULL,
  method = "NB",
  ncore = 1,
  cell.lim = 10
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{cluster.ctrl}{Select the clusters that we want to detect DEGs, it/they as 
the control.}

\item{cluster.sam}{Select the clusters that we want to detect DEGs, it/they as 
the sample.}

\item{cell.ctrl}{Select the cells as the reference cells for detecting DEGs.}

\item{cell.sam}{Select the cells as the sample cells for detecting DEGs.}

\item{frac}{A fraction cutoff, the cluster marker genes expressed at least a 
cutoff fraction of the cluster cells.}

\item{logFC}{The cutoff of log Fold-change for differentially expressed marker 
genes.}

\item{Padj}{The cutoff of the adjusted P-value. If Padj is NULL, use p-value 
< 0.05 as a threshold. Set Padj as 1, without any cutoff.}

\item{latent.factor}{The latent factor from coldata, which represents number 
values or factors, and only one latent factor can be inputed.}

\item{method}{Which method is used to identify cluster markers, two options: 
'NB' for Negative Binomial model and 'QP' for QuasiPoisson model.}

\item{ncore}{The multiple cores for parallel calculating.}

\item{cell.lim}{The minimum cells for each cluster to calculate marker genes.}
}
\description{
This is the basic function in RISC, it can identify the differentially expressed 
genes (DEGs) by comparing samples between the selected clusters. The criteria 
used for the cluster markers are also appropriate to DEGs.
}
\details{
Here RISC provides two algorithms to detect DEGs, the primary one is a model 
"Quasi-Poisson" which has advantage to identify DEGs from the cluster with 
a small number of cells. Meanwhile, RISC also has alternative algorithm: 
"Negative Binomial" model.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
obj0 = scPCA(obj0, npc = 10)
obj0 = scUMAP(obj0, npc = 3)
obj0 = scCluster(obj0, slot = "cell.umap", k = 3, method = 'density')
DimPlot(obj0, slot = "cell.umap", colFactor = 'Cluster', size = 2)
DEG0 = scDEG(obj0, cluster.ctrl = 1, cluster.sam = 3, 
       cell.lim = 3, method = 'QP')
}
\references{
Paternoster et al., Criminology (1997)

Berk et al., Journal of Quantitative Criminology (2008)

Liu et al., Nature Biotech. (2021)
}
