% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{scDisperse}
\alias{scDisperse}
\title{Processing Data.}
\usage{
scDisperse(
  object,
  method = "loess",
  min.UMI = NULL,
  mean.cut = NULL,
  QP_bin = 100,
  lspan = 0.05,
  top.var = NULL,
  pval = 0.5
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{method}{What method is used to define dispersion, now support "QP" and "loess", 
the default method is "loess".}

\item{min.UMI}{A cutoff of the minimum UMIs of each gene, the genes expression 
below the min.UMI will be discarded from highly variable genes. The default value
is 100 when input NULL.}

\item{mean.cut}{A cutoff of the average value of each gene, the genes expression 
outside the mean.cut range will be discarded from highly variable genes. The 
input is a range vector, like c(0.1, 5).}

\item{QP_bin}{The number of fragments using to fit dispersion in the Quasi-Poinsson 
model, how many bins are formed in the regression.}

\item{lspan}{The number of parameter using to fit dispersion, controlling the degree 
of smoothing in the loess model.}

\item{top.var}{The maximum number of highly variable genes, the default is NULL, 
including all the highly variable genes.}

\item{pval}{The P-value is used to cut off the highly variable genes, 
the default is 0.5.}
}
\value{
RISC single cell dataset, the metadata slot.
}
\description{
After data scaling, RISC will identify highly variably expressed genes, based on 
Quasi-Poinsson model, where the coefficient of variation is calculated for each 
gene (C > 0.5 as a cutoff for highly variable genes). Then, to controlling for 
the relationship between S (standard deviation) and mean (average value), 
Quasi-Poisson regression is used to further filter the genes with over-
dispersion C caused by small mean. Lastly, RISC estimates the corresponding 
ratio with r between the observed C and the predicted C of each gene, with a 
threshold r > 1.
}
\examples{
# RISC object
obj0 = raw.mat[[5]]
obj0 = scFilter(obj0, min.UMI = 0, max.UMI = Inf, min.gene = 10, min.cell = 3)
obj0 = scNormalize(obj0)
obj0 = scDisperse(obj0)
}
\references{
Liu et al., Nature Biotech. (2021)
}
