% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{scFilter}
\alias{scFilter}
\title{The Pre-processing Data.}
\usage{
scFilter(
  object,
  min.UMI = NULL,
  max.UMI = NULL,
  min.gene = NULL,
  min.cell = NULL,
  mitochon = 1,
  gene.ratio = 0.05,
  is.filter = TRUE
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{min.UMI}{The min UMI for valid cells is usually based on the distribution 
analysis, with more than 2.5 percentage of UMI distribution, discarding the 
cells with too few UMIs. This parameter can be adjusted manually.}

\item{max.UMI}{The max UMI for valid cells is usually based on the distribution 
analysis, with less than 97.5 percentage of UMI distribution, discarding the 
cells with too many UMIs.This parameter can be adjusted manually.}

\item{min.gene}{The min number of expressed genes for valid cells. The default 
is based on the distribution analysis, with more than 0.5 percentage of gene 
distribution. This parameter can be adjusted manually.}

\item{min.cell}{The min number of cells for valid expressed genes. The default 
is based on the distribution analysis, with more than 0.5 percentage of cell 
distribution. This parameter can be adjusted manually.}

\item{mitochon}{The cutoff of the mitochondrial UMI proportion for valid cells.}

\item{gene.ratio}{The cutoff of the proportions of genes in UMIs.}

\item{is.filter}{Whether filter the data.}
}
\value{
RISC single cell dataset, the coldata and rowdata slots.
}
\description{
After input data, RISC preliminarily filter datasets by using three criteria: 
first, discard cells with too low/high raw counts/UMIs by distribution analysis. 
Second, remove cells with too low expressed genes. Lastly, filter out the genes 
only expressed in few cells.
}
\examples{
# RISC object
obj0 = raw.mat[[5]]
obj0 = scFilter(obj0, min.UMI = 0, max.UMI = Inf, min.gene = 10, min.cell = 3)
}
\references{
Liu et al., Nature Biotech. (2021)
}
