% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{FilterPlot}
\alias{FilterPlot}
\title{Processing Plot}
\usage{
FilterPlot(object, colFactor = NULL)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{colFactor}{Use the factor (column name) in the coldata to make a processing 
plot, but each time only one column name can be inputted.}
}
\description{
The "FilterPlot" function makes the plots to show the UMIs and expressed genes 
of individual cells. These plots are usually used to estimate the data before 
and after pre-processing, so the users can visually select the optimal 
parameters to filter data.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
FilterPlot(obj0, colFactor = 'Group')
}
\references{
Wickham, H. (2016)

Auguie, B. (2015)

Liu et al., Nature Biotech. (2021)
}
