% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{read10X_h5}
\alias{read10X_h5}
\title{Import data from 10X Genomics output (h5).}
\usage{
read10X_h5(file.path, is.filter = TRUE)
}
\arguments{
\item{file.path}{The path of the filtered 10X Genomics output (h5 file).}

\item{is.filter}{Remove not expressed genes.}
}
\value{
RISC single cell dataset, including count, coldata, and rowdata.
}
\description{
Import data directly from 10X Genomics output, usually using filtered gene 
matrices which contains h5 file. 
The user only need to input the directory into "data.path". If not the original 
10X Genomics output, the user can use 'readsc' function.
}
