% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{read10X_mtx}
\alias{read10X_mtx}
\title{Import data from 10X Genomics output (tsv-mtx).}
\usage{
read10X_mtx(data.path, sep = "\\t", is.filter = TRUE)
}
\arguments{
\item{data.path}{Directory containing the filtered 10X Genomics output, 
including three files: matrix.mtx, barcode.tsv (without colnames) and gene.tsv 
(without colnames).}

\item{sep}{The sep can be changed by the users}

\item{is.filter}{Remove not expressed genes.}
}
\value{
RISC single cell dataset, including count, coldata, and rowdata.
}
\description{
Import data directly from 10X Genomics output, usually using filtered gene 
matrices which contains three files: matrix.mtx, barcode.tsv and gene.tsv. 
The user only need to input the directory into "data.path". If not the original 
10X Genomics output, the user have to make sure the barcode.tsv and gene.tsv 
without col.names, the barcode.tsv at least contains one column for cell 
barcode, and the gene.tsv has two columns for gene Ensembl ID and Symbol.
}
