% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{readHTSeqdata}
\alias{readHTSeqdata}
\title{Import data from HT-Seq output.}
\usage{
readHTSeqdata(data.path, is.filter = TRUE)
}
\arguments{
\item{data.path}{Directory containing all the HT-Seq outputs, each HT-Seq has the
same length of genes.}

\item{is.filter}{Remove not expressed genes.}
}
\value{
RISC single cell dataset, including count, coldata, and rowdata.
}
\description{
Import data directly from HT-Seq output, but each HTSeq.output.txt need to have 
the same length of genes. If genes annotated by Ensembl ID, the customer need to 
transfer them to gene symbols manually. The user need to make a folder to contain 
all HT-Seq outputs, and input the directory into "data.path".
}
