% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{readsc}
\alias{readsc}
\title{Import data from matrix, cell and genes directly.}
\usage{
readsc(count, cell, gene, is.filter = TRUE)
}
\arguments{
\item{count}{Matrix with raw counts/UMIs.}

\item{cell}{Data.frame with cell Barcode, whose row.name are equal to the 
col.name of the matrix.}

\item{gene}{Data.frame with gene symbol, whose row.name are the same as the 
row.name of the matrix.}

\item{is.filter}{Remove not expressed genes.}
}
\value{
RISC single cell dataset, including count, coldata, and rowdata.
}
\description{
Import data set from matrix, cell and genes directly, the customer needs three 
files: a matrix file including gene expression values: raw counts/UMIs (rows for 
genes while columns for cells), a cell file (whose row.name are equal to the 
col.name of the matrix), and a gene file whose row.name are the same as the 
row.name of the matrix. If row.names of the gene matrix are Ensembl ID, the 
customer need to transfer them to gene symbols manually.
}
\examples{
mat0 = as.matrix(raw.mat[[1]])
coldata0 = as.data.frame(raw.mat[[2]])
coldata.obj = coldata0[coldata0$Batch0 == 'Batch3',]
matrix.obj = mat0[,rownames(coldata.obj)]
obj0 = readsc(count = matrix.obj, cell = coldata.obj, 
       gene = data.frame(Symbol = rownames(matrix.obj), 
       row.names = rownames(matrix.obj)), is.filter = FALSE)
}
