% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrating.R
\name{InPlot}
\alias{InPlot}
\title{Integration Plot}
\usage{
InPlot(
  object = NULL,
  var.gene = NULL,
  Colors = NULL,
  nPC = 20,
  neighbor = 30,
  res = 1,
  method = "louvain",
  algorithm = "kd_tree",
  ncore = 1,
  minPC = 11,
  Std.cut = 0.95,
  bin = 5
)
}
\arguments{
\item{object}{A list of RISC objects.}

\item{var.gene}{The highly variable genes.}

\item{Colors}{The colors labeling for different data sets.}

\item{nPC}{The PCs will be calculated.}

\item{neighbor}{The nearest neighbors.}

\item{res}{The resolution of cluster searched for, works in "louvain" method.}

\item{method}{The method of cell clustering for individual datasets.}

\item{algorithm}{The algorithm for knn, the default is "kd_tree", all options: 
"kd_tree", "cover_tree", "CR", "brute".}

\item{ncore}{The number of multiple cores for testing.}

\item{minPC}{The minimal PCs for detecting cell clustering.}

\item{Std.cut}{The cutoff of standard deviation of the PCs.}

\item{bin}{The bin number for calculating cell clustering.}
}
\description{
The "InPlot" function makes the plot to show how the PCs explain the variance 
for data integration. This plot helps the users to select the optimal reference 
and the PCs to perform data integration.
}
\references{
Liu et al., Nature Biotech. (2021)
}
