% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrating.R
\name{scLargeIntegrate}
\alias{scLargeIntegrate}
\title{Integrating Multiple Large Datasets}
\usage{
scLargeIntegrate(
  objects,
  K = 10,
  npc = 50,
  var.gene = NULL,
  Add.Id = NULL,
  ncore = 1,
  seed = 123
)
}
\arguments{
\item{objects}{The list of multiple RISC objects: 
list{object1, object2, object3, ...}.}

\item{K}{The number of PCs used for data integration, the optimal number of 
PCs is equal to the number of PCs used for dimension reduction of individual 
datasets.}

\item{npc}{The number of the PCs returns from "scMultiIntegrate" function, 
they are usually used for the subsequent analyses, like cell embedding.}

\item{var.gene}{Define the variable genes mannually. Here input a vector of gene 
names as variable genes}

\item{Add.Id}{Add a vector of Ids to label different datasets, a character vector.}

\item{ncore}{The number of multiple cores for data integration.}

\item{seed}{The random seed to keep consistent result.}
}
\description{
The "scMultiIntegrate" function can be used for data integration of multiple 
datasets, it is basically based on our new algorithm: reference principal 
components integration (RPCI). RPCI decomposes all the target datasets based 
on the reference. The output of this function can be used for low dimension 
visualization.
}
\references{
Liu et al., Nature Biotech. (2021)
}
