% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrating.R
\name{MSC}
\alias{MSC}
\title{Alignment of gene expression values}
\usage{
MSC(X, Y)
}
\arguments{
\item{X}{The reference matrix, row for genes and column for cells.}

\item{Y}{The target matrix, row for genes and column for cells.}
}
\description{
This funciton is not used in data integration but for adjusting the results, 
usually the users do not need this.
}
\examples{
## Calculate MSC with the inputs of two matrices
mat0 = as.matrix(raw.mat[[1]])
coldata0 = as.data.frame(raw.mat[[2]])
coldata1 = coldata0[coldata0$Batch0 == 'Batch1',]
coldata2 = coldata0[coldata0$Batch0 == 'Batch4',]
mat1 = mat0[,rownames(coldata1)]
mat2 = mat0[,rownames(coldata2)]
msc0 = MSC(mat1, mat2)
}
\references{
Mevik et al., JSS (2007)
}
