% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrating.R
\name{scMultiIntegrate}
\alias{scMultiIntegrate}
\title{Integrating Multiple Datasets}
\usage{
scMultiIntegrate(
  objects,
  eigens = 10,
  add.Id = NULL,
  var.gene = NULL,
  method = "RPCI",
  align = "OLS",
  npc = 50,
  adjust = TRUE,
  ncore = 1,
  do.fast = "AUTO",
  seed = 123
)
}
\arguments{
\item{objects}{The list of multiple RISC objects: 
list{object1, object2, object3, ...}. The first set is the reference to generate 
gene-eigenvectors.}

\item{eigens}{The number of eigenvectors used for data integration.}

\item{add.Id}{Add a vector of Id to label different datasets, a character vector.}

\item{var.gene}{Define the variable genes manually. Here input a vector of gene 
names as variable genes}

\item{method}{Two methods are available, RPCI algorithm and SIMPLS algorithm, 
the default is RPCI.}

\item{align}{The method for alignment of gene expression values: "Optimal" for 
alignment by experience, "Predict" for alignment by RPCI prediction, and "OLS" 
for alignment by the ordinary linear regression.}

\item{npc}{The number of the PCs returns from "scMultiIntegrate" function, 
they are usually used for the subsequent analyses, like cell embedding and 
cell clustering.}

\item{adjust}{Whether adjust the number of eigenvectors.}

\item{ncore}{The number of multiple cores for data integration.}

\item{do.fast}{With enough PC memory, we could do.fast, otherwise, 
let it as the default. Three options: AUTO (default), TRUE, and FALSE.}

\item{seed}{The random seed to keep consistent result.}
}
\description{
The "scMultiIntegrate" function can be used for data integration of multiple 
datasets, it is basically based on our new approach RPCI (reference principal 
components integration), which decomposes all the target datasets based on the 
reference data. The output of this function is RISC object, including the 
integrated eigenvectors and aligned gene expression values.
}
\examples{
data("raw.mat")
obj1 = raw.mat[[3]]
obj2 = raw.mat[[4]]
obj0 = list(obj1, obj2)
var0 = intersect(obj1@vargene, obj2@vargene)
obj0 = scMultiIntegrate(obj0, eigens = 10, var.gene = var0, 
                       npc = 15, add.Id = c("Set-1", "Set-2"))
obj0 = scUMAP(obj0, npc = 10, use = "PLS")
DimPlot(obj0, slot = "cell.umap", colFactor = "Set", size = 2)
DimPlot(obj0, slot = "cell.umap", colFactor = "Group", size = 2, label = TRUE)
}
\references{
Liu et al., Nature Biotech. (2021)
}
