% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{scNormalize}
\alias{scNormalize}
\title{The Processing Data.}
\usage{
scNormalize(
  object,
  method = "robust",
  libsize = 1e+06,
  remove.mito = FALSE,
  norm.dis = TRUE,
  large = FALSE,
  ncore = 1
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{method}{A method for scdataset normalization, two options: "cosine" and 
"robust".}

\item{libsize}{The standard sum of the UMI in each cell.}

\item{remove.mito}{Remove mitochondrial genes from library size.}

\item{norm.dis}{Normalize the distribution of count data.}

\item{large}{Whether a large size data (ncell > 50,000)}

\item{ncore}{The multiple cores for parallel calculating.}
}
\value{
RISC single cell dataset, the assay and rowdata slots.
}
\description{
After data filtration, RISC will normalized the raw counts/UMIs by using size 
factors which are calculated by the raw counts/UMIs of each cell and will 
remove sequencing depth batch. The output will be transformed in log1p. The gene 
expression values of RISC object for the subsequent analyses is based on the 
normalized data. Here two kinds of normalization can be employed: one is based on 
the least absolute deviations, while the other is from the least square.
}
\examples{
# RISC object
obj0 = raw.mat[[5]]
obj0 = scFilter(obj0, min.UMI = 0, max.UMI = Inf, min.gene = 10, min.cell = 3)
obj0 = scNormalize(obj0)
}
\references{
Boscovich, R.J. (1757)

Thompson, W.J., Computers in Physics (1992)
}
