% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_Dimension.R
\name{scPCA}
\alias{scPCA}
\title{Dimension Reduction.}
\usage{
scPCA(object, npc = 20)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{npc}{The number of PCs will be generated based on highly variable genes 
(usually < 1,500), npc equal to the first 20 PCs as the default.}
}
\value{
RISC single cell dataset, the DimReduction slot.
}
\description{
Based on highly variably expressed genes of the datasets, RISC calculates the 
principal components (PCs) of the cells using prcomp functions. The major PCs, 
which explain most gene expression variance, are used for dimension reduciton.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
obj0 = scPCA(obj0, npc = 10)
}
\references{
Jolliffe et al. (2016)

Alter et al., PNAS (2000)

Gonzalez et al., JSS (2008)

Mevik et al., JSS (2007)
}
