% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{PCPlot}
\alias{PCPlot}
\title{Processing Plot}
\usage{
PCPlot(object)
}
\arguments{
\item{object}{RISC object: a framework dataset.}
}
\description{
The "PCPlot" function makes the plot to show How the PCs explain the variance. 
This plot helps the users to select the optimal PCs to perform dimension 
reduction and data integration.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
obj0 = scPCA(obj0, npc = 10)
PCPlot(obj0)
}
\references{
Wickham, H. (2016)

Auguie, B. (2015)

Liu et al., Nature Biotech. (2021)
}
