% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{raw.mat}
\alias{raw.mat}
\title{Example data}
\format{
A list including a simulated cell-gene matrix, columns for cells and 
rows for genes, a cell group and a batch information.
}
\usage{
data(raw.mat)
}
\description{
Example data
}
\keyword{datasets}
