% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{scScale}
\alias{scScale}
\title{Processing Data.}
\usage{
scScale(object, method = "scale", center = TRUE, scale = FALSE)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{method}{A model used for scale scdataset, the default is root-mean-square 
scaling.}

\item{center}{Whether to center the matrix.}

\item{scale}{Whether using standard deviation to scale the matrix.}
}
\value{
The scaled matrix.
}
\description{
After data normalization, RISC will perform root-mean-square scaling to the 
dataset and generated scaled counts which balance expression levels in each cell 
with empirical mean equal to 0. Therefore, only biological signal will be 
reserved in scaled counts. RISC utilizes scaled counts for dimension reduction 
and data integration.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
scale.mat = scScale(obj0)
}
\references{
Juszczak et al., CiteSeer (2002)

Jiawei et al. (2011)
}
