% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{class}
\name{RISCdata-class}
\alias{RISCdata-class}
\title{RISC data}
\value{
RISC object: a S4 framework dataset
}
\description{
The RISC object contains all the basic information used in single cell RNA-seq 
analysis, including raw counts/UMIs, normalized gene values, dimension reduction, 
cell clustering, and so on. The framework of RISC object is a S4 dataset, 
consisting of assay, coldata, rowdata, metadata, vargene, cluster, and 
DimReduction.
}
\section{Slots}{

\describe{
\item{\code{assay}}{The list of gene counts/UMIs: raw and normalized counts}

\item{\code{coldata}}{The data.frame with cell information, such as cell types, stages, 
and other factors.}

\item{\code{rowdata}}{The data.frame with gene information, such as coding or non-coding 
genes.}

\item{\code{metadata}}{The data.frame with meta value.}

\item{\code{vargene}}{The highly variable gene. 
These genes are utilized in dimension reduction.}

\item{\code{cluster}}{The cell clustering information: include three algorithms for 
clustering.}

\item{\code{DimReduction}}{The values of dimension reduction.}
}}

