% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{RISCdata}
\alias{RISCdata}
\alias{.RISC_show}
\alias{RISC}
\alias{object}
\title{RISC data}
\usage{
.RISC_show(object)
}
\arguments{
\item{object}{RISC object: a S4 framework dataset}
}
\description{
This will show the full information of RISC object, including the number of cells, 
the number of genes, any biological or statistical information of cells or genes.
}
