% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SingleCellData}
\alias{SingleCellData}
\title{Import single cell data}
\usage{
SingleCellData(assay, coldata, rowdata)
}
\arguments{
\item{assay}{The list of gene counts.}

\item{coldata}{The data.frame with cell information.}

\item{rowdata}{The data.frame with gene information.}
}
\value{
SingleCellData
}
\description{
The single cell RNA-seq (scRNA-seq) data can be imported in three different ways.
Primarily, we could import from 10X Genomics output directly by using 
"read10Xgenomics". The user only need to provide the folder path. Secondly, 
we could read data from HT-seq output by "readHTSeqdata", the user have to input 
the folder path. Lastly, we could input matrix, cell and genes mannually 
using "readscdata".
}
