% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{SubSet}
\alias{SubSet}
\title{Utilities Subset data}
\usage{
SubSet(object, cells = NULL, genes = NULL)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{cells}{The cells are directly used for collecting a data subset.}

\item{genes}{The genes are directly used for collecting a data subset.}
}
\description{
The "Subset" function can abstract a data subset from the full dataset, this 
function not only collect the subset of coldata and rowdata, but also abstract 
raw counts/UMIs. Meanwhile, after "Subset" function, RISC object need to be 
normalized and scaled one more time.
}
\examples{
# RISC object
obj0 = raw.mat[[5]]
obj0
cell1 = rownames(obj0@coldata)[1:15]
obj1 = SubSet(obj0, cells = cell1)
obj1
}
