% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce_Dimension.R
\name{scTSNE}
\alias{scTSNE}
\title{Dimension Reduction.}
\usage{
scTSNE(
  object,
  npc = 20,
  embedding = 2,
  use = "PCA",
  perplexity = 30,
  seed = 123
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{npc}{The number of PCs (or PLS) using for t-SNE, the default is 20, 
but need to be modified by the users. The PCA for individual dataset, while 
PLS for the integrated data.}

\item{embedding}{The number of components t-SNE output.}

\item{use}{What components used for t-SNE: PCA or PLS.}

\item{perplexity}{Perplexity parameter: if the cell numbers are small, 
decrease this parameter, otherwise tSNE cannot be calculated.}

\item{seed}{The random seed to keep tSNE result consistent.}
}
\value{
RISC single cell dataset, the DimReduction slot.
}
\description{
The t-SNE is calculated based on the eigenvectors of single cell dataset, and 
the user can select the eigenvectors manually. Of note, the selected eigenvectors 
directly affect t-SNE values. 
For the integrated data (the result of "scMultiIntegrate" funciton), RISC utilizes
the PCR output "PLS" to calculate the t-SNE, therefore, the user has to input 
"PLS" in "use = ", instead of the defaut parameter "PCA".
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
obj0 = scPCA(obj0, npc = 10)
obj0 = scTSNE(obj0, npc = 4, perplexity = 10)
DimPlot(obj0, slot = "cell.tsne", colFactor = 'Group', size = 2)
}
\references{
Laurens van der Maaten, JMLR (2014)
}
