% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{UMAPlot}
\alias{UMAPlot}
\title{UMAP Plots}
\usage{
UMAPlot(
  object,
  colFactor = NULL,
  genes = NULL,
  legend = TRUE,
  Colors = NULL,
  size = 0.5,
  Alpha = 0.8,
  plot.ncol = NULL,
  raw.count = FALSE,
  exp.range = NULL,
  exp.col = "firebrick2"
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{colFactor}{Use the factor (column name) in the coldata to make a UMAP plot, 
but each time only one column name can be inputted.}

\item{genes}{Use the gene expression values (gene symbol) to make UMAP plots, 
each time more than one genes can be inputted.}

\item{legend}{Whether a legend shown at UMAP plot.}

\item{Colors}{The users can use their own colors (color vector). The default of 
the "UMAPlot" funciton will assign colors automatically.}

\item{size}{Choose the size of dots at UMAP plots, the default size is 0.5.}

\item{Alpha}{Whether show transparency of individual points, the default is 0.8.}

\item{plot.ncol}{If the users input more than one genes, the arrangement of 
multiple UMAP plots depends on this parameter.}

\item{raw.count}{If use normalized or raw counts.}

\item{exp.range}{The gene expression cutoff for plot, e.g. "c(0, 1.5)" for 
expression level between 0 and 1.5.}

\item{exp.col}{The gradient color for gene expression.}
}
\description{
The UMAP plots are widespread in scRNA-seq data analysis. Here, the "UMAPlot" 
function not only can make plots for factor labels of individual cells but also 
can show gene expression values of each cell.
}
\references{
Wickham, H. (2016)

Auguie, B. (2015)
}
