% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{ViolinPlot}
\alias{ViolinPlot}
\title{Violin Plot}
\usage{
ViolinPlot(
  object,
  colFactor = NULL,
  genes = NULL,
  legend = TRUE,
  trim = TRUE,
  Colors = NULL,
  Alpha = 0.8,
  dots = TRUE,
  wid = "area"
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{colFactor}{Use the factor (column name) in the coldata to make heatmap, but 
each time only one column name can be inputted.}

\item{genes}{The gene expression pattern: gene symbol}

\item{legend}{Whether a legend shown at heatmap.}

\item{trim}{Whether trim the violin plot}

\item{Colors}{The users can use their own colors (color vector). The default of 
the "UMAPlot" funciton will assign colors automatically.}

\item{Alpha}{Whether show transparency of individual points, the default is 0.8.}

\item{dots}{Adding jitter dots to the violin plot. The default is TRUE.}

\item{wid}{The scale format, options: "area", "width", "count". 
The default: "area"}
}
\description{
The "ViolinPlot" map makes plots to show gene expression patterns of the clusters 
or other factors. The default groups cells into the clusters, but the users can 
input the factor (column name) of coldata.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
ViolinPlot(obj0, colFactor = 'Group', genes = 'Gene718')
}
\references{
Wickham, H. (2016)

Auguie, B. (2015)
}
