% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMarker.R
\name{AllMarker}
\alias{AllMarker}
\title{Find All Cluster Markers}
\usage{
AllMarker(
  object,
  positive = TRUE,
  frac = 0.25,
  log2FC = 0.5,
  Padj = 0.05,
  latent.factor = NULL,
  min.cells = 25L,
  method = "QP",
  ncore = 1
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{positive}{Whether only output the cluster markers with positive log2FC.}

\item{frac}{A fraction cutoff, the marker genes expressed at least a 
cutoff fraction of all the cells.}

\item{log2FC}{The cutoff of log2 Fold-change for differentially expressed marker 
genes.}

\item{Padj}{The cutoff of the adjusted P-value.}

\item{latent.factor}{The latent factor from coldata, which represents number 
values or factors, and only one latent factor can be inputed.}

\item{min.cells}{The threshold for the cell number of valid clusters.}

\item{method}{Which method is used to identify cluster markers, three options: 'NB' 
for Negative Binomial model, 'QP' for QuasiPoisson model, and 'Wilcox' for Wilcoxon 
Rank Sum and Signed Rank model.}

\item{ncore}{The multiple cores for parallel calculating.}
}
\description{
This function depends on "scMarker" function by using the same criteria, and 
generates markers for all the clusters. Here, if the cell number of any cluster 
is less than 10 , RISC will skip this cluster and not detect its cluster markers 
in the default parameters.
}
\details{
Because log2 cannot handle counts with value 0, we use log1p to calculate average 
values of counts and log2 to format fold-change.
}
