% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{scDisperse}
\alias{scDisperse}
\title{Processing Data.}
\usage{
scDisperse(
  object,
  method = "quasipoisson",
  mean.cut = NULL,
  ncut = 20,
  top.var = NULL,
  FDR = 0.1,
  ratio = 1.25
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{method}{What method is used to define dispersion: quasipoisson or residual.
The default method is "Quasi-Poisson".}

\item{mean.cut}{A cutoff of the average value of each gene, the genes expression 
outside the mean.cut range will be discarded from highly variable genes. The 
input is a range vector, like c(0.1, 5).}

\item{ncut}{The number of fragments using to fit dispersion. The Quasi-Poinsson 
model divides genes into a number of bins based on the gene expression. This 
parameter indicated how many bins are used for Quasi-Poinsson model.}

\item{top.var}{The maximum number of highly variable genes, the default is NULL, 
including all the highly variable genes.}

\item{FDR}{The cutoff for variable genes.}

\item{ratio}{To change mean.cut by ratio, and make RISC robustly to select 
variable genes. The user can use the default value.}
}
\value{
RISC single cell dataset, the metadata slot.
}
\description{
After data scaling, RISC will identify highly variably expressed genes, based on 
Quasi-Poinsson model, where the coefficient of variation is calculated for each 
gene (C > 0.5 as a cutoff for highly variable genes). Then, to controlling for 
the relationship between S (standard deviation) and mean (average value), 
Quasi-Poisson regression is used to further filter the genes with over-
dispersion C caused by small mean. Lastly, RISC estimates the corresponding 
ratio with r between the observed C and the predicted C of each gene, with a 
threshold r > 1.
}
\examples{
# RISC object
obj0 = raw.mat[[5]]
obj0 = scFilter(obj0, min.UMI = 0, max.UMI = Inf, min.gene = 10, min.cell = 3)
obj0 = scNormalize(obj0)
obj0 = scDisperse(obj0)
}
\references{
Liu et al., Nature Biotech. (2021)
}
