% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{Heat}
\alias{Heat}
\title{Heatmap}
\usage{
Heat(
  object,
  colFactor = NULL,
  genes = NULL,
  cells = NULL,
  gene.lab = FALSE,
  gene.cluster = 0,
  sample_bin = FALSE,
  ann_col = NULL,
  lim = NULL,
  smooth = "smooth",
  span = 0.75,
  degree = 1,
  palette = NULL,
  num = 50,
  con.bin = TRUE,
  cell.lab.size = 10,
  gene.lab.size = 5,
  value_only = FALSE,
  ...
)
}
\arguments{
\item{object}{RISC object: a framework dataset.}

\item{colFactor}{Use the factor (column name) in the coldata to make heatmap, but 
be factors.}

\item{genes}{Use the gene expression values (gene symbol) to make heatmap, need to 
be inputted by the users.}

\item{cells}{Use the subset cells of the whole coldata (cells) to make heatmap, 
the default is NULL and including all the cells.}

\item{gene.lab}{Whether label gene names for the heatmap.}

\item{gene.cluster}{The cluster numbers for gene clustering in the heatmap. 
The default is 0, without clustering genes.}

\item{sample_bin}{The cell aggregating in samples, the default is FALSE.}

\item{ann_col}{The annotation colors for colFactors, the input is a list.}

\item{lim}{The gene expression range shown at heat-maps.}

\item{smooth}{If use smooth to adjust heatmap, the default is "smooth" and another
choice is "loess".}

\item{span}{The loess span.}

\item{degree}{The loess degree.}

\item{palette}{The color palette used for heatmap. The default is 
brewer.pal(n = 7, name = "RdYlBu").}

\item{num}{The cells for individual bin spans.}

\item{con.bin}{Whether use consistent bin span.}

\item{cell.lab.size}{The font size for column.}

\item{gene.lab.size}{The font size for row.}

\item{value_only}{Only return values.}
}
\description{
The "Heat" map makes heatmap to show gene expression patterns of single cells. 
The default groups cells into clusters, so the column of heatmap represents 
genes while the row of heatmap for the clusters of all the cells.
}
\examples{
# RISC object
obj0 = raw.mat[[3]]
gene0 = c('Gene718', 'Gene120', 'Gene313', 'Gene157', 'Gene30', 
          'Gene325', 'Gene415', 'Gene566', 'Gene990', 'Gene13', 
          'Gene604', 'Gene934', 'Gene231', 'Gene782', 'Gene10')
Heat(obj0, colFactor = 'Group', genes = gene0, gene.lab = TRUE, gene.cluster = 3, 
sample_bin = TRUE, lim = 2, gene.lab.size = 8)
}
\references{
Kolde, R. (2015)
}
