% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrating.R
\name{MSC}
\alias{MSC}
\title{Alignment of gene expression values}
\usage{
MSC(X, Y)
}
\arguments{
\item{X}{The reference matrix, row for genes and column for cells.}

\item{Y}{The target matrix, row for genes and column for cells.}
}
\description{
This funciton is not used in data integration but for adjusting the results, 
usually the users do not need this.
}
\references{
Mevik et al., JSS (2007)
}
